+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 - add: sort by creation time
 - add: randomly select
 - fix: fix some bugs
+- add: download list window
 # v0.2.0
 - improve: AddonNodeView
 - fix: WorkshopVpkAddon.FullVpkFilePath is sometimes not updated